Move definition of TypeLocation into its own header
